types：refactor：切り分け
