+++ src/components/elements/CalendarEl.vue
@@ -1,4 +1,11 @@
 <script setup lang="ts">
+import type {
+  CategoryEntry,
+  DateLogs,
+  LogEntry,
+  SubCategory,
+  SubCategoryEntry,
+} from '@/common/types';
 import categoryData from 'samples/features/data/category.json';
 import subCategoryData from 'samples/features/data/subcategories.json';
 import { Calendar as VCalendar } from 'v-calendar';
@@ -6,34 +13,6 @@ import type { CalendarDay } from 'v-calendar/dist/types/src/utils/page.js';
 import 'v-calendar/style.css';
 import { type Ref, ref, watch } from 'vue';
 import CalendarWatch from './CalendarWatch.vue';
-
-export type LogEntry = {
-  category_id: number;
-  subcategory_id: number;
-  price: number;
-};
-export type DateLogs = {
-  id: number;
-  timestamp: string;
-  registered_date: string;
-  logs: LogEntry[];
-};
-export type CategoryEntry = {
-  id: number;
-  category: string;
-};
-export type Category = {
-  id: number;
-  category: CategoryEntry[];
-};
-export type SubCategoryEntry = {
-  id: number;
-  category: string;
-};
-export type SubCategory = {
-  id: number;
-  subcategories: SubCategoryEntry[];
-};
 
 // eslint-disable-next-line @typescript-eslint/typedef
 const props = defineProps<{
